requirements.txt
opencv_python>=4.9.0.80
Pillow>=9.5.0
ultralytics>=8.0.47
